+++ src/HoYo/Bookmark.hs
@@ -8,16 +8,19 @@ module HoYo.Bookmark (
   , BookmarkSearchTerm (..)
   , formatBookmark
   , formatBookmarks
+  , DefaultBookmark (..)
 
   -- *  Working with bookmarks
   , searchBookmarks
   , filterBookmarks
   , filterBookmarkByName
   , filterBookmarkByDirInfix
+  , bookmarksFromDefault
 
   -- ** Parsing\/encoding bookmarks from\/to TOML
   , bookmarkCodec
   , bookmarksCodec
+  , defaultBookmarkCodec
   , decodeBookmarks
   , decodeBookmarksFile
   , encodeBookmarks
@@ -33,8 +36,10 @@ import Data.Function
 import Data.List
 import qualified Data.Text as T
 
-import Control.Monad (void)
+import Control.Monad (forM, void)
 import Control.Monad.IO.Class
+
+import Data.Time
 
 import qualified Toml
 import Toml (TomlCodec)
@@ -48,6 +53,12 @@ bookmarkCodec = Bookmark
   <*> Toml.int        "index"               .== _bookmarkIndex
   <*> Toml.zonedTime  "created"             .== _bookmarkCreationTime
   <*> Toml.dioptional (Toml.string "name")  .== _bookmarkName
+
+-- | A 'TomlCodec' for encoding and decoding 'Bookmark's.
+defaultBookmarkCodec :: TomlCodec DefaultBookmark
+defaultBookmarkCodec = DefaultBookmark
+  <$> Toml.string     "directory"           .== _defaultBookmarkDirectory
+  <*> Toml.dioptional (Toml.string "name")  .== _defaultBookmarkName
 
 -- | A 'TomlCodec' for encoding and decoding 'Bookmarks'.
 bookmarksCodec :: TomlCodec Bookmarks
@@ -97,3 +108,12 @@ filterBookmarks :: Maybe String -> Maybe String -> Bookmark -> Bool
 filterBookmarks name dirInfix = combAnd
                                     (filterBookmarkByName name)
                                     (filterBookmarkByDirInfix dirInfix)
+
+-- | Convert a list of 'DefaultBookmark's to a 'Bookmarks', assiging indices and/
+-- creation times on the fly.
+bookmarksFromDefault :: MonadIO m => [DefaultBookmark] -> m Bookmarks
+bookmarksFromDefault dbms = Bookmarks <$> bms
+  where
+    bms = forM (zip dbms [1..]) $ \(DefaultBookmark dir name, idx) -> do
+      zTime <- liftIO getZonedTime
+      return $ Bookmark dir idx zTime name