+++ docs/CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to the
 
 ### Added
 
+- Added an array table "default_bookmark" to config: these bookmarks
+will be added to the bookmark list on init/clear.
+
 ### Fixed
 
 ### Changed